Mostly-working dhis2 maps CRA2 POC
