--- .babelrc
@@ -1,10 +1,0 @@
-{
-    "presets": ["react", "es2015", "stage-0"],
-    "plugins": ["transform-object-assign"],
-    "env": {
-        "test": {
-            "presets": ["react", "es2015", "stage-0"],
-            "plugins": ["transform-es2015-modules-commonjs"]
-        }
-    }
-}
