+++ .eslintignore
@@ -1,4 +1,3 @@
 node_modules/
 build/
-webpack.config.js
 *.spec.js
