+++ .eslintrc.json
@@ -1,30 +1,3 @@
 {
-    "extends": ["eslint:recommended", "plugin:react/recommended", "prettier"],
-    "env": {
-        "browser": true,
-        "node": true,
-        "es6": true
-    },
-    "parser": "babel-eslint",
-    "parserOptions": {
-        "ecmaFeatures": {
-            "classes": true,
-            "modules": true,
-            "jsx": true
-        }
-    },
-    "plugins": ["prettier"],
-    "rules": {
-        "prettier/prettier": "error"
-    },
-    "settings": {
-        "react": {
-            "version": "detect"
-        }
-    },
-    "globals": {
-        "DHIS_CONFIG": true,
-        "L": true,
-        "ee": true
-    }
+    "extends": ["react-app"]
 }
