+++ cypress.json
@@ -1,4 +1,4 @@
 {
-    "baseUrl": "http://localhost:8082",
+    "baseUrl": "http://localhost:3000",
     "projectId": "uqdb76"
 }
