--- index.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" dir="ltr">
-
-<head>
-    <meta charset="UTF-8">
-    <title>DHIS2 Maps</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link href="../favicon.ico" rel="shortcut icon">
-    <link href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700" rel="stylesheet">
-    <!-- jss-insertion-point -->
-</head>
-
-<body>
-    <div id="app"></div>
-    <div id="context-menu"></div>
-
-    <%= htmlWebpackPlugin.options.vendorScripts %>
-    <%= htmlWebpackPlugin.options.serviceWorkerScript %>
-</body>
-
-</html>
