+++ public/favorites.html
