+++ public/i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2019-02-04T16:30:15.843Z\n"
-"PO-Revision-Date: 2019-02-04T16:30:15.843Z\n"
+"POT-Creation-Date: 2018-12-05T09:55:15.234Z\n"
+"PO-Revision-Date: 2018-12-05T09:55:15.234Z\n"
 
 msgid "Maps"
 msgstr ""
