+++ src/components/map/BoundaryLayer.js
@@ -1,3 +1,5 @@
+/* global L */
+
 import i18n from '@dhis2/d2-i18n';
 import Layer from './Layer';
 import { filterData } from '../../util/filter';
