+++ src/components/map/Map.js
@@ -1,3 +1,5 @@
+/* global L */
+
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
