+++ src/components/map/TrackedEntityLayer.js
@@ -1,3 +1,5 @@
+/* global L */
+
 import i18n from '@dhis2/d2-i18n';
 import { apiFetch } from '../../util/api';
 import Layer from './Layer';
