+++ src/epics/earthEngine.js
@@ -1,3 +1,5 @@
+/* global ee */
+
 import i18n from '@dhis2/d2-i18n';
 import { combineEpics } from 'redux-observable';
 import 'rxjs/add/operator/concatMap';
