+++ src/reducers/map.js
@@ -260,11 +260,6 @@ const map = (state = defaultState, action) => {
                 ],
             };
 
-        case types.LAYER_ADD_DATA:
-            return {
-                ...state,
-            };
-
         case types.LAYER_REMOVE:
             return {
                 ...state,
@@ -295,7 +290,6 @@ const map = (state = defaultState, action) => {
         case types.ORGANISATION_UNIT_SELECT:
         case types.ORGANISATION_UNIT_UNSELECT:
         case types.ORGANISATION_UNIT_COORDINATE_CHANGE:
-        case types.ORGANISATION_UNITS_FILTER:
         case types.DATA_FILTER_SET:
         case types.DATA_FILTER_CLEAR:
         case types.MAP_EARTH_ENGINE_VALUE_SHOW:
