+++ src/util/requests.js
@@ -97,7 +97,7 @@ const upgradeGisAppLayers = layers => {
 export const getUrlParameter = name => {
     name = name.replace(/[[]/, '\\[').replace(/[\]]/, '\\]');
     const regex = new RegExp('[\\?&]' + name + '=([^&#]*)');
-    const results = regex.exec(location.search);
+    const results = regex.exec(window.location.search);
     return results === null
         ? ''
         : decodeURIComponent(results[1].replace(/\+/g, ' '));
